Scalar fix for codegenmodule jinja template
All parameters were being treated as matrices in the template code that copies values into the parameter structure. For scalar parameters, this was causing the parameter value itself to be passed in as the first argument to memcpy instead of a pointer. Added an if clause to handle scalar parameters separately.